requirements.txt
openai
finnhub-python
yfinance
python-dotenv
